Update README.md
The THRUSTfund contract offers a secure and innovative way to manage the time-locking of ERC20 tokens, such as DeltaV THRUST and others. It is designed with features that allow for the detailed management of token locks, auto-sending tokens to beneficiaries, and manual withdrawals under specific conditions. This platform is tailored for token holders looking for a reliable solution to lock tokens over time while managing gas fees efficiently. Ideal for those new to the blockchain space, it provides a straightforward approach to token management and investment growth.